[Inline] Fix incorrectly dropped noalias metadata
This is the same fix as 5800478e8f03085391ece5c033dc7b9cec6bc841,
just for CloneScopedAliasMetadata rather than PropagateCallSiteMetadata.

In this case the previous outcome was incorrectly dropped metadata,
as it was not part of the computed metadata map.

The real change in the test is that the first load now retains
metadata, the rest of the changes are due to changes in metadata
numbering.