[AArch64] Enable Cortex-A55 schedmodel
The model was committed in 8a5f82c19cd6f75254672baf548aee9ba0c3d006
but not yet enabled to allow for a few fix ups. This adds a few
of these fixes, and also a LLVM MCA test to check most instructions.
While I do have plans to look into some more tuning, it's time to
enable this as it better than using the A53 schedule.

Differential Revision: https://reviews.llvm.org/D88017