[NFC] Fix -Wsometimes-uninitialized
After 40649f03e93e468df8908be7edbcd4a04c03fa73, clang detects that MUL may be uninitialized. Set it to nullptr to suppress this check.

Adding an assert to check that it is ultimately set fails two test cases. Since this is not a new issue, leave the assertion commented out until a code owner can fix the bug. The two failing test cases are noted in the assertion comment.