[ARM] Don't handle low overhead branches in AnalyzeBranch
It turns our that the BranchFolder and IfCvt does not like unanalyzable
branches that fall-through. This means that removing the unconditional
branches from the end of tail predicated instruction can run into
asserts and verifier issues.

This effectively reverts 5e866598f35a4ee2b8019b691aba9121e1aec230, but
adds handling to t2DoLoopEndDec which are not branches, so can be safely
skipped.