Reland "[PDB] Defer relocating .debug$S until commit time and parallelize it"
This reverts commit 7571397e397d9268d68a0784fa0905e0ee3a6d85 and relands
f873a194937065b865e84c55d395b0cac733b5e1.

The ASan error was debugged and determined to be the fault of an invalid
object file input in our test suite, which was fixed by my last change.
LLD's project policy is that it assumes input objects are valid, so I
have added a comment about this assumption to the relocation bounds
check.