Recommit "[RISCV] Legalize select when Zbt extension available"
This recommits 1cb44066977c6ef9f1cbbf2808ec0a747b2343bd with
the polarity of some of the pattern corrected.

Original commit message:
The custom expansion of select operations in the RISC-V backend
interferes with the matching of cmov instructions. Legalizing
select when the Zbt extension is available solves that problem.

Reviewed By: luismarques, craig.topper

Differential Revision: https://reviews.llvm.org/D93767